[DEV-6383] Move misplaced calc_deployment_metrics_line_github test
